First go at File Upload
This isn't complete - there's no way to download yet, and what the
interface should be for actually streaming out content isn't clear to
me. I'm also not totally sure that this persister addition method is the
right way to do things. However, I think it gets things in a place where we can
begin to iterate based on use cases.