Custom predicates with Fedora resources, ref #198
Valkyrie was making up predicates when persisting attributes to Fedora.
We needed a way to provide a predicate for each attribute in the
resource.

Fedora's metadata adapter will accept a hash as a schema and use any
defined predicate for a given term. If no predicate is defined, the
current default one will be used.

Custom schemas must be provided when initializing the metadata adapter
in config/initializers/valkyrie.rb.