Removing useless local variable assignment
Prior to this commit, Rubocop identified a useless assignment. I believe
the intention of the prior code was to set the `file_size` property of
the Resource. However, as written, the assignment was "useless" because
it set a local variable `file_size` (instead of the property).

I believe the fix is to prepend `self.` to the `file_size` assignment,
thus setting the expected property instead of the local variable.

```ruby
def size=(size)
  file_size = size
end
```

Related to d5528472f91d89617e64b1ef4185410a6be0b204