Add single valued support for Solr/In-Memory
The in-memory adapter has no need for casting to/from multiple values.

The strategy for Solr and the other adapters is to cast single-valued
arrays to a single value for Dry::Types to deal with. Set and Array will
cast it back appropriately if that's what the property is set to be.

Closes #461
Closes #462