Randomizing order of spec execution
Prior to this commit, it appeared that the specs ran in a sequential
order. Assuming they are running in the same order during each test,
this could mean that we have unknown temporal dependencies in our code.

Could we switch Valkyrie's specs to run in random order, thus helping
tease out if there exist temporal dependencies in our code.

_In running them randomly it appears to preserve meaningful
documentation format._

Closes #797