Tidying up spec documentation output
Prior to this commit, the documentation read as follows:

```console
Valkyrie::Types
  #<Dry::Types[Constructor<Nominal<Valkyrie::Persistence::OptimisticLockToken> fn=lib/valkyrie/types.rb:57>]>
    casts from a string
```

With the change, we now get:

```console
Valkyrie::Types
  Valkyrie::Types::OptimisticLockToken
    casts from a string
```

I find this more legible and less surprising when reading the specs.
With the prior format, I'm left wondering "is there a stray puts call?"
I see the prior output as something "programmatic and as not clear as
documentation."

This attempts to tidy that up. However, I see this as something up for
discussion.