Making sure Valkyrie uses DateTime consistently.
Valkyrie normalizes DateTime values to UTC to ensure consistency.
This commit changes the method used from `.utc` to `.new_offset(0)`.
The former returns a Time rather than a DateTime, which causes
the type checking in the strict DateTime type to fail. It also
rewrites tests to expect the DateTime format for the UTC zone,
'+00:00' rather than 'UTC'.