Use brackets instead of super in Config methods
We first noticed issues in Ruby 3 trying to do

```rb
Valkyrie::Config.new.resource_class_resolver
```

It returned `nil` even though in this case the default resource class
resolver was defined. The same problem appears to exist for other
Config methods. Changing them to use `self[:method_name]` instead of
`super` circumvents the issue, which appears to be a bug in OpenStruct,
probably with Ruby 3.