☑️ Remove ChangeSet#prepopulate! test
This test is fragile in that it does not account for the fact that we
can specify a :prepopulate option for a property.

When you specify the `:prepopulate` option (as in
`Hyrax::Form::PcdmCollection`) and use this shared spec, then the now
deleted spec fails with no recourse.

Related to:

- https://github.com/samvera/hyrax/issues/6702
- https://github.com/samvera/hyrax/pull/6703